+++ R/didor.R
@@ -14,6 +14,8 @@
 #' @importFrom magrittr %>%
 #' @importFrom rlang .data abort %||%
 #' @importFrom purrr pmap map_chr
+#' @importFrom cli cli_alert_info
+#' @importFrom lubridate ymd_hms with_tz minutes format_ISO8601 now
 ## usethis namespace: end
 NULL