+++ R/get-data.R
@@ -1,21 +1,27 @@
-#' Get data
+#' @name get_data
 #'
-#' Get the data of the last millesime of all the datafiles found in data. All
-#' columns are returned as `chr` (see below).
+#' @title Get data
 #'
-#' You can use `convert()` to convert number and integer.
+#' @description Get the data of the last millesime of all the datafiles found in data. All columns are returned as `chr` (see below). \cr
 #'
 #' For private life reason, data returned by DiDo can be secretize (the value is
-#' replaced by the string "secret") so readr can't determine data type.
+#' replaced by the string "secret") so readr can't determine data type. You can
+#' use `convert()` to convert number and integer.
+#'
+#' get_data cache data before loading it. By default it saves the files
+#' in `tempdir()`. If you downloaded the same data again, il will first try to
+#' find it in the `cache` argument. If you want to cache data between
+#' session, don't keep the default but use your own directory.
+#'
 #'
 #' @param data a tibble issued from `datafiles() or a dataframe with two columns
 #'   `rid` and `millesime`.
 #' @param query a query to pass to the API to select columns and filter on
 #'   values.
-#' @param concat `TRUE`
+#' @param concat `TRUE` if `TRUE`, returns a tibble with all data concatenated in one tibble, else returns a list of tibbles.
 #' @param col_types how to convert columns, the default is to use char for all
 #'   columns `cols(.default = "c")`
-#' @param directory the directory to cache/save downloaded files. Default is `tempdir()`
+#' @param cache the directory to cache/save downloaded files. Default is `tempdir()`
 #'
 #' @return
 #'
@@ -23,7 +29,14 @@
 #' one tibble.
 #' If concat is `FALSE`, return a list of tibbles.
 #'
-#' `get_data()` returns only chr
+#' `get_data()` returns only chr columns. Use `convert` to convert columns to good types.
+#'
+#' @details
+#'
+#' For caching, `get_data()` will use a reproductible name compose of the
+#' datafile identifier (rid) and the stringification of the query passed to
+#' `get_data()` of a query is passed.
+#'
 #'
 #' @export
 #'
@@ -42,12 +55,12 @@
 #'   get_data(query = c(DEPARTEMENT_CODE = "eq:971"))
 #' datafiles() %>%
 #'   dido_search("drom") %>%
-#'   get_data(query = c(DEPARTEMENT_CODE = "eq:971"), directory = tempdir())
+#'   get_data(query = c(DEPARTEMENT_CODE = "eq:971"), cache = tempdir())
 get_data <- function(data,
                      query = list(),
                      col_types = cols(.default = "c"),
                      concat = TRUE,
-                     directory = tempdir()) {
+                     cache = tempdir()) {
   if (missing(data)) {
     stop("argument is mandatory.")
   }
@@ -65,7 +78,7 @@ get_data <- function(data,
     }
   }
 
-  dir.create(directory, recursive = TRUE, showWarnings = FALSE)
+  dir.create(cache, recursive = TRUE, showWarnings = FALSE)
 
   if (!"millesime" %in% names(data)) {
     mill <- last_millesime(data)
@@ -78,8 +91,8 @@ get_data <- function(data,
     stop("argument include more than one millesime per datafile")
   }
 
-  millesime_keys <- select(mill, "rid", "millesime")
-  list_df <- pmap(millesime_keys, ~ get_csv(..1, ..2, query, col_types, directory = directory))
+  millesime_keys <- mill %>% left_join(datafiles(), by = "rid") %>% select("rid", "millesime", "last_modified")
+  list_df <- pmap(millesime_keys, ~ get_csv(..1, ..2, ..3, query, col_types, cache = cache))
 
   columns <- millesimes(millesime_keys) %>%
     columns(quiet = TRUE)